PBS: Make state saving more robust
Recent Android versions are sending a onStop signal to the mediasession
receiver when the Notification is removed or swiped away. This caused
the PBS to set index of the playing song to the begining which made it
lose the actual index the user stopped.

This commit changes that by reacting with pause on that signature which
preserves the index.

Signed-off-by: Hendrik Borghorst <hendrikborghorst@gmail.com>